Word Choices, Spelling fixes.
Pre-Moderate, is a better concept to make it understandable that content is not pre moderated or checked. However, it can be removed etc.
changed supercede to supersede - 
takedown is not one word, rather two words, I added a hyphen. 

Cancelled or Canceled  I guess this will turn into a debate, but from my research cancelled, overtook canceled in the 1980's, although many Americans, see Canceled as good enough.