Remove indemnification for frivolous lawsuits
A user should not be liable to Github simply because someone files a lawsuit against Github that in some way relates to the user's conduct.  For example, let's say I create an app called, "Empowering Women," which a totally fictitious man we'll call "Donald Trump" finds offensive.  Mr. Trump then sues Github for negligent infliction of emotional distress.  The suit is frivolous and will be rapidly dismissed, but these terms before my modification would have required *me* to pay Github's attorney's fees.

The modification I've made allows for indemnification only if Github is sued because of the user's violation of the terms of service, not for any lawsuit, no matter how frivolous, and no matter how tangentially related, to the user's conduct on Github.  If Github would like more protection than that, it should purchase an insurance policy, not foist liability on its users.