Completed the fork definition
Copied from:

1. https://help.github.com/articles/dmca-takedown-policy/#b-what-about-forks-or-whats-a-fork

The definition of what a GitHub fork means. Currently the ToS only states that fork is a bare copy of a GitHub repository. However as defined on the referenced article, a fork means also the users can make changes to their copy, instead of get them as read-only copies.

Related issues:

1. https://github.com/github/site-policy/issues/7 OSS licences vs. ToS: grants of rights
1. https://github.com/github/site-policy/issues/37 The term "fork" is not defined anywhere in the GitHub Terms of Service
1. https://github.com/github/site-policy/issues/38 More permissive license grant to other users