Update to Privacy Statement for Spring 2018
Over the last few months, we've gotten a few questions asking about our General Data Protection Regulation (GDPR) compliance. We are proud to announce that we are compliant with the GDPR. Additionally, we have always provided the same level of privacy protection to our users regardless of their residency, location, or citizenship, and that will not change. We provide strong privacy and security protection to _all_ of our users.

For the most part, our changes to the Privacy Statement are only points of clarification. GitHub doesn't ask for more personal data from our users than we need to provide our services to you. Where we offer you the option of giving us more data, we provide you the ability to access and delete the data you have given us. For example, you can always remove your profile information, your comments in issues, and your repository contents. We have gone through our Privacy Statement to provide more context and transparency, though, so our users understand exactly why we ask for information and what we'll do with it.

### GDPR Compliance

* The GDPR requires us to inform our users about the legal basis on which we process their data. In this update, we explain what data we collect and why.
* We describe our security practices in more detail.
* We now provide a separate page describing our tracking, our use of cookies, and listing our subprocessors (the vendors and third parties we have engaged to process personal data on our behalf).
* Throughout the Privacy Statement, we provide greater transparency and insight into our data collection, data handling, data retention, and data deletion processes.
* If you are a Corporate Terms of Service customer and you need a Data Protection Agreement with us, please [contact support](https://github.com/contact). We will be happy to provide one. Please understand that with the GDPR compliance deadline coming up, our volume of requests is high, but we will respond to you as promptly as possible.