Create github-subprocessors-and-cookies.md
This page provides a list of GitHub's subprocessors, such as our vendors and service providers. It also offers some transparency into the cookies we set and why we set them, as well as describing exactly where we do tracking for analytics purposes on our site and who our analytics providers are (at the moment, we use Google Analytics, but if we bring on additional providers, we will be able to provide fuller transparency here).