Mark Fernandes's improvements to the main GitHub terms and conditions
Please contact me on mark.j.fernandes@gmail.com if you disagree with any of my changes or you are unsure about any of them.

I use the general rule of thumb that less is better.

No need to say things like 'thank you'. If you want to say 'thank you', say it on your home page. People have little time, and want to spend as little time as possible reading user agreements such as this one.

Sometimes I've changed the ordering of statements because it makes for easier comprehension.

Sometimes I've just taken text out that is unnecessary.

I've tried to make changes so that for the vast majority of your users, it is quicker to get through them. Eg. majority of users probably don't use the APIs, so I've changed the terms so that you aren't forced to read the terms relating to the APIs if it is not relevant to you.

On the face of it, it may seem better to inform users about everything so that they are fully informed. However, in reality, users (especially ones not representing organisations), find it hard to mentally keep in mind all of the details of a user agreement. This means that informing them about everything can actually mean that they are less aware rather than more aware. If you try to be more selective with what you communicate, they can end-up being in fact more informed because they find the communication less unwieldy.

Using bullet points can make for better comprehension because it is generally a better visual organisation of information.

Sometimes, I've moved text so that a 'you can skip the rest of this section' statement can be made. So what happens, is that you read important stuff at the top (text sometimes moved to the top), and then you hit such a skipping statement so that you can skip the rest of the section.

Sometimes sentences have been combined to say the same thing with less words, in hopefully a clearer way. On the other hand, at least in one place I've split a very long-winded sentence into a few sentences, to make for easier comprehension. You find such very long-winded sentences in legal contracts. Even though they may be legally sound, they are generally very difficult to read.

Most of the changes should be easy to understand.

Further Advice
==============

You can:
- look at the advice I gave to Microsoft regarding one of their terms and conditions documents, that can be accessed by going to: https://answers.microsoft.com/en-us/skype/forum/all/processing-the-microsoft-services-agreement-legal/e104804e-c2c0-413f-b8e5-180a2b029415 .
- look at how the BBC website terms and conditions (https://www.bbc.co.uk/usingthebbc/terms/) are arranged, and try to draw inspiration from them, in order to simplify the presentation, and improve the accessibility of, your terms.