Lowercase;spacing;clearer
Headings for collapsible text blocks are now in lower-case rather than upper-case. The reason for this, is that previously the headings appeared to stick out too much, as though they were major headings in the document (when they really weren't).

An extra blank line has been placed just before every new main section in the agreement, to make it clearer where new sections begin.

One sentence has been re-written, so that it is simpler and shorter.