Improved Definitions section
Mostly by using indentation and vertically listed bullet points, I've tried to make the definitions easier to read.

Space and ink isn't much at a premium when viewing the agreement on a computer. Therefore, we can be more liberal with space and ink than we would be under the scenario of having to solely optimise the agreement for use in hard-copy form.