Knowing -> needing information
The act of knowing is a personal human act.

When processing a user agreement, technically speaking (especially with advances in AI) knowing is often not quite the right term to use in cases where users need information. A user may obtain information on something (for example, for cancelling their account), without actually needing to know that information.

With the above in mind, relevant passages that made reference to 'knowing' have been accordingly changed to instead reference the acquisition of information.