Update github-acceptable-use-policies.md
**tl;dr: here's our take on a first iteration of changes based on your feedback, key takeaways:** 
- **dual use security research content remains explicitly welcome on GitHub**
- **by “harm” we mean “causing technical harms, such as overconsumption of resources, physical damage, downtime, denial of service, or data loss”**
- **we removed any disclosure methodology bias based on status of vulnerability knowledge**
- **we clarified what we mean by “restriction” (authentication-based where possible) and under which circumstances it might apply**
- **we clarified the intent of our security contact recommendation and made it more flexible**

**:eyes:  at the [changes](https://github.com/github/site-policy/pull/397/files) and read below for more details**

With these updates, we want to reiterate that this effort and conversation is an attempt to narrow restrictions in our existing policies which we felt were too broad in scope, and to explicitly allow dual use software to exist on the platform. We feel it is crucial to the future of software security that developers and security researchers are able to collaborate freely on our platform.

With such explicit allowances comes a need to capture those corner cases where we need to be able to respond to platform abuse. Specifically, these are the cases where GitHub, the platform, is abused to deliver malware and exploits into victim infrastructure. The challenge is to broadly welcome dual use security research content on GitHub, but also make it explicit that using GitHub as a malware or exploit CDN in support of unlawful activity is not allowed.

Our intent is to capture this by tying incident response to specific abuse reports we receive from others. We do not want to restrict or qualify dual use security research content, but rather whether or not something is being hosted on the platform with the purpose of facilitating unlawful activity. The actionable difference would be a project whose majority interactions are from victim infrastructure as evidenced by abuse reports.

This requires clear, actionable language that guides our incident response teams, but that is also sufficiently narrow in scope so that we err on the side of open and free research and development. 

Any action we take should only affect a specific instance of a project tied to a specific abuse, in collaboration with the project owner where possible (i.e. when we are able to establish contact), and with the lightest touch required to halt the abuse (e.g. by making the content accessible only to authenticated users in the case of automated malware CDN abuse).

We understand that GitHub is not going to solve decades of passionate debate on the subject of dual use security research, but we felt it was time to have this conversation and reaffirm that security research has a stable home on GitHub. 

We look forward to your continued contributions and comments as we continue to iterate based on your feedback. If the community strongly prefers to leave our [AUP and CG](https://docs.github.com/en/github/site-policy/github-community-guidelines) as it currently exists in place, we are also open to that feedback.