Update github-community-guidelines.md
Thanks very much for the continued contributions to this PR. As we continue to listen and iterate based on community feedback, we’ve incorporated many of your code review suggestions in this latest set of revisions to our proposed updates to the Acceptable Use Policies (AUP) and Community Guidelines, to:

- More clearly and narrowly define the scope of platform abuse
- Move examples of abusive content (such as using the platform to manage command and control servers) back to the Acceptable Use Policies
- Broadly and explicitly exempt dual use technology and elaborate on this exemption early on in our Community Guidelines
- Remove redundancy in the elaboration of the policy in the Community Guidelines
- Move examples of what we mean by technical harms to the Community Guidelines
- Clearly indicate in which cases of platform abuse restrictions might apply, and what form those restrictions take
- Further clarify that our SECURITY.md contact recommendation is not a requirement

:eyes: at the changes https://github.com/github/site-policy/pull/397/files and read below for more details.

For a full summary of previous iterations and updates based on your feedback so far, please see the [opening PR comment](https://github.com/github/site-policy/pull/397#issue-626295321), which we've updated with that history.
***
Again, the goal of these updates is to remove any overly broad restrictions on dual use technology on GitHub as it exists in our current policy, and to provide clear guidelines for both ourselves and the security community as a whole that enable, welcome and encourage security research and collaboration on our platform.

As we draw closer to the end of our 30-day comment period on June 1, 2021, we invite your continued discussion and feedback on these changes. If you have direct updates to the AUP or Community Guidelines language you’d like to propose, we strongly encourage the use of commit suggestions in your PR comments.

We would like to thank the community members, project maintainers, and developers who have shared feedback with us in the PR and have reached out for live discussions on this topic. Your feedback and suggestions have been tremendously valuable throughout this process. :sparkles: