Spring 2022 updates to Privacy Statement
In this pull request we’re proposing changes to our Privacy Statement to help make our policies clearer and easier to understand.

Here are the key updates:
- Refreshed language to clarify our privacy practices and reflect requirements of new and updated privacy laws. These changes are not meant to change or narrow any of our previous privacy commitments. For example, in order to reduce redundancies and focus on personal data as a general term, we removed references to User Personal Information.
- You now have more granular control over display of your data in connection with Community Features. You can check out the new private profile setting under “Contributions & Activity” here.
- You can now find our California Privacy Notice by simply scrolling to the end of the Privacy Statement.

Note that while you may see a lot of green (new) content in the diff, in some cases, this is because we’ve moved content around to convey our privacy practices in a simpler way. It doesn’t necessarily mean that we’ve changed the language. For example, we included a new section “Who is responsible for the processing of your information?” that contains some information from the “Organizations” section of our previous version.

These updates will go into effect after the 30-day notice and comment period, on May 28, 2022 at 3pm PT.